backend/requirements.txt
@"
fastapi==0.104.1
uvicorn==0.24.0
groq==0.4.2
python-dotenv==1.0.0
reportlab==4.0.4
pydantic==2.4.2
python-multipart==0.0.6
"@ | Out-File -FilePath backend\requirements.txt -Encoding UTF8